+++ pmd-apex/src/main/java/net/sourceforge/pmd/lang/apex/rule/performance/AvoidSoqlInLoopsRule.java
@@ -4,18 +4,13 @@
 
 package net.sourceforge.pmd.lang.apex.rule.performance;
 
-import net.sourceforge.pmd.lang.apex.ast.ASTBlockStatement;
-import net.sourceforge.pmd.lang.apex.ast.ASTDoLoopStatement;
-import net.sourceforge.pmd.lang.apex.ast.ASTForEachStatement;
-import net.sourceforge.pmd.lang.apex.ast.ASTForLoopStatement;
-import net.sourceforge.pmd.lang.apex.ast.ASTReturnStatement;
 import net.sourceforge.pmd.lang.apex.ast.ASTSoqlExpression;
-import net.sourceforge.pmd.lang.apex.ast.ASTWhileLoopStatement;
-import net.sourceforge.pmd.lang.apex.rule.AbstractApexRule;
-import net.sourceforge.pmd.lang.ast.Node;
 
-public class AvoidSoqlInLoopsRule extends AbstractApexRule {
-
+/**
+ * @deprecated use {@link OperationWithLimitsInLoopRule}
+ */
+@Deprecated
+public class AvoidSoqlInLoopsRule extends AbstractAvoidNodeInLoopsRule {
     public AvoidSoqlInLoopsRule() {
         setProperty(CODECLIMATE_CATEGORIES, "Performance");
         // Note: Often more complicated as just moving the SOQL a few lines.
@@ -26,31 +21,6 @@ public class AvoidSoqlInLoopsRule extends AbstractApexRule {
 
     @Override
     public Object visit(ASTSoqlExpression node, Object data) {
-        if (insideLoop(node) && parentNotReturn(node)) {
-            addViolation(data, node);
-        }
-        return data;
-    }
-
-    private boolean parentNotReturn(ASTSoqlExpression node) {
-        return !(node.getParent() instanceof ASTReturnStatement);
-    }
-
-    private boolean insideLoop(ASTSoqlExpression node) {
-        Node n = node.getParent();
-
-        while (n != null) {
-            if (n instanceof ASTBlockStatement && n.getParent() instanceof ASTForEachStatement) {
-                // only consider the block of the for-each statement, not the iterator
-                return true;
-            }
-            if (n instanceof ASTDoLoopStatement || n instanceof ASTWhileLoopStatement
-                    || n instanceof ASTForLoopStatement) {
-                return true;
-            }
-            n = n.getParent();
-        }
-
-        return false;
+        return checkForViolation(node, data);
     }
 }